+++ src/server/game/Battlegrounds/Zones/BattlegroundAV.cpp
@@ -141,7 +141,7 @@ void BattlegroundAV::HandleQuestComplete(uint32 questid, Player* player)
     if (GetStatus() != STATUS_IN_PROGRESS)
         return;//maybe we should log this, cause this must be a cheater or a big bug
     TeamId teamId = player->GetTeamId();
-    uint8 CFSteamId =GetOtherTeamId(player->GetTeamId());
+    uint8 CFSteamId = GetOtherTeamId(player->GetTeamId());
     //TODO add reputation, events (including quest not available anymore, next quest availabe, go/npc de/spawning)and maybe honor
 #if defined(ENABLE_EXTRAS) && defined(ENABLE_EXTRA_LOGS)
     sLog->outDebug(LOG_FILTER_BATTLEGROUND, "BG_AV Quest %i completed", questid);
