+++ src/server/game/Handlers/CharacterHandler.cpp
@@ -1220,6 +1220,7 @@ void WorldSession::HandlePlayerLoginFromDB(LoginQueryHolder* holder)
 
     sScriptMgr->OnPlayerLogin(pCurrChar);
     delete holder;
+    
     if (pCurrChar->GetTeamId() != pCurrChar->GetCFSTeamId())
         pCurrChar->FitPlayerInTeam(pCurrChar->GetBattleground() && !pCurrChar->GetBattleground()->isArena() ? true : false, pCurrChar->GetBattleground());
 }
