+++ src/server/game/Handlers/QueryHandler.cpp
@@ -36,7 +36,7 @@ void WorldSession::SendNameQueryOpcode(uint64 guid)
     data << playerData->name;                       // played name
     data << uint8(0);                               // realm name - only set for cross realm interaction (such as Battlegrounds)
     Player* player = ObjectAccessor::FindPlayerInOrOutOfWorld(guid);
-    data << uint8(player ? player->getRace() :playerData->race);
+    data << uint8(player ? player->getRace() : playerData->race);
     data << uint8(playerData->gender);
     data << uint8(playerData->playerClass);
